TrackInfo / MediaSamplesBlock should be in their own headers https://bugs.webkit.org/show_bug.cgi?id=299086 rdar://160854758
Reviewed by Eric Carlson.

No change in obvservable behaviours.

* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/MediaSample.h:
(WebCore::TrackInfo::isAudio const): Deleted.
(WebCore::TrackInfo::isVideo const): Deleted.
(WebCore::TrackInfo::type const): Deleted.
(WebCore::TrackInfo::operator== const): Deleted.
(WebCore::TrackInfo::TrackInfo): Deleted.
(WebCore::VideoInfo::create): Deleted.
(WebCore::VideoInfo::VideoInfo): Deleted.
(WebCore::AudioInfo::create): Deleted.
(WebCore::AudioInfo::AudioInfo): Deleted.
(WebCore::MediaSamplesBlock::MediaSampleItem::isSync const): Deleted.
(WebCore::MediaSamplesBlock::MediaSamplesBlock): Deleted.
(WebCore::MediaSamplesBlock::setInfo): Deleted.
(WebCore::MediaSamplesBlock::info const): Deleted.
(WebCore::MediaSamplesBlock::protectedInfo const): Deleted.
(WebCore::MediaSamplesBlock::presentationTime const): Deleted.
(WebCore::MediaSamplesBlock::duration const): Deleted.
(WebCore::MediaSamplesBlock::presentationEndTime const): Deleted.
(WebCore::MediaSamplesBlock::isSync const): Deleted.
(WebCore::MediaSamplesBlock::trackID const): Deleted.
(WebCore::MediaSamplesBlock::isVideo const): Deleted.
(WebCore::MediaSamplesBlock::isAudio const): Deleted.
(WebCore::MediaSamplesBlock::type const): Deleted.
(WebCore::MediaSamplesBlock::append): Deleted.
(WebCore::MediaSamplesBlock::size const): Deleted.
(WebCore::MediaSamplesBlock::isEmpty const): Deleted.
(WebCore::MediaSamplesBlock::clear): Deleted.
(WebCore::MediaSamplesBlock::takeSamples): Deleted.
(WebCore::MediaSamplesBlock::discontinuity const): Deleted.
(WebCore::MediaSamplesBlock::setDiscontinuity): Deleted.
(isType): Deleted.
* Source/WebCore/platform/MediaSamplesBlock.h: Added.
(WebCore::MediaSamplesBlock::MediaSampleItem::isSync const):
(WebCore::MediaSamplesBlock::MediaSamplesBlock):
(WebCore::MediaSamplesBlock::setInfo):
(WebCore::MediaSamplesBlock::info const):
(WebCore::MediaSamplesBlock::protectedInfo const):
(WebCore::MediaSamplesBlock::presentationTime const):
(WebCore::MediaSamplesBlock::duration const):
(WebCore::MediaSamplesBlock::presentationEndTime const):
(WebCore::MediaSamplesBlock::isSync const):
(WebCore::MediaSamplesBlock::trackID const):
(WebCore::MediaSamplesBlock::isVideo const):
(WebCore::MediaSamplesBlock::isAudio const):
(WebCore::MediaSamplesBlock::type const):
(WebCore::MediaSamplesBlock::append):
(WebCore::MediaSamplesBlock::size const):
(WebCore::MediaSamplesBlock::isEmpty const):
(WebCore::MediaSamplesBlock::clear):
(WebCore::MediaSamplesBlock::takeSamples):
(WebCore::MediaSamplesBlock::discontinuity const):
(WebCore::MediaSamplesBlock::setDiscontinuity):
* Source/WebCore/platform/TrackInfo.cpp: Copied from Source/WebCore/platform/graphics/avfoundation/FormatDescriptionUtilities.h.
(WebCore::convertEnumerationToString):
* Source/WebCore/platform/TrackInfo.h: Added.
(WebCore::TrackInfo::isAudio const):
(WebCore::TrackInfo::isVideo const):
(WebCore::TrackInfo::type const):
(WebCore::TrackInfo::operator== const):
(WebCore::TrackInfo::toVariant const):
(WebCore::TrackInfo::TrackInfo):
(WebCore::VideoInfo::create):
(WebCore::VideoInfo::VideoInfo):
(WebCore::AudioInfo::create):
(WebCore::AudioInfo::AudioInfo):
(isType):
(WTF::LogArgument<WebCore::TrackInfoTrackType>::toString):
* Source/WebCore/platform/audio/cocoa/AudioDecoderCocoa.cpp:
* Source/WebCore/platform/graphics/AudioVideoRenderer.h:
* Source/WebCore/platform/graphics/MediaPlayer.cpp:
* Source/WebCore/platform/graphics/MediaPlayerEnums.h:
(WTF::LogArgument<WebCore::TrackInfoTrackType>::toString): Deleted.
* Source/WebCore/platform/graphics/SourceBufferPrivate.cpp:
* Source/WebCore/platform/graphics/SourceBufferPrivate.h:
* Source/WebCore/platform/graphics/avfoundation/AudioVideoRendererAVFObjC.h:
* Source/WebCore/platform/graphics/avfoundation/FormatDescriptionUtilities.h:
* Source/WebCore/platform/graphics/cocoa/AudioTrackPrivateWebM.cpp:
* Source/WebCore/platform/graphics/cocoa/CMUtilities.h:
* Source/WebCore/platform/graphics/cocoa/CMUtilities.mm:
* Source/WebCore/platform/graphics/cocoa/H264UtilitiesCocoa.mm:
* Source/WebCore/platform/graphics/cocoa/SourceBufferParserWebM.h:
* Source/WebCore/platform/graphics/cocoa/VP9UtilitiesCocoa.mm:
* Source/WebCore/platform/graphics/cocoa/WebMAudioUtilitiesCocoa.mm:
* Source/WebCore/platform/mediarecorder/MediaRecorderPrivateEncoder.cpp:
* Source/WebCore/platform/mediarecorder/MediaRecorderPrivateWriter.cpp:
* Source/WebCore/platform/mock/mediasource/MockSourceBufferPrivate.cpp:
* Source/WebKit/GPUProcess/media/RemoteTrackInfo.h:
* Source/WebKit/Scripts/webkit/messages.py:
(headers_for_type):
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:

Canonical link: https://commits.webkit.org/300164@main
