+++ Source/WebCore/platform/MediaSample.h
@@ -26,11 +26,7 @@
 #pragma once
 
 #include <WebCore/FloatSize.h>
-#include <WebCore/FourCC.h>
-#include <WebCore/HdrMetadataType.h>
 #include <WebCore/MediaPlayerEnums.h>
-#include <WebCore/PlatformVideoColorSpace.h>
-#include <WebCore/SharedBuffer.h>
 #include <functional>
 #include <variant>
 #include <wtf/MediaTime.h>
@@ -48,11 +44,8 @@ typedef const struct opaqueCMFormatDescription *CMFormatDescriptionRef;
 
 namespace WebCore {
 
-class FragmentedSharedBuffer;
 class MockSampleBox;
 class ProcessIdentity;
-class SharedBuffer;
-struct TrackInfo;
 
 using TrackID = uint64_t;
 
@@ -161,157 +154,7 @@ public:
     }
 };
 
-struct AudioInfo;
-struct VideoInfo;
-
-struct TrackInfo : public ThreadSafeRefCounted<TrackInfo> {
-    using TrackType = TrackInfoTrackType;
-
-    bool isAudio() const { return type() == TrackType::Audio; }
-    bool isVideo() const { return type() == TrackType::Video; }
-
-    TrackType type() const { return m_type; }
-
-    bool operator==(const TrackInfo& other) const
-    {
-        if (type() != other.type() || codecName != other.codecName || trackID != other.trackID)
-            return false;
-        return equalTo(other);
-    }
-
-    FourCC codecName;
-    String codecString;
-    TrackID trackID { 0 };
-
-    virtual ~TrackInfo() = default;
-
-protected:
-    virtual bool equalTo(const TrackInfo& other) const = 0;
-    TrackInfo(TrackType type)
-        : m_type(type) { }
-
-private:
-    const TrackType m_type { TrackType::Unknown };
-};
-
-struct VideoInfo : public TrackInfo {
-    static Ref<VideoInfo> create() { return adoptRef(*new VideoInfo()); }
-
-    FloatSize size;
-    // Size in pixels at which the video is rendered. This is after it has
-    // been scaled by its aspect ratio.
-    FloatSize displaySize;
-    uint8_t bitDepth { 8 };
-    PlatformVideoColorSpace colorSpace;
-
-    RefPtr<SharedBuffer> atomData;
-
-private:
-    VideoInfo()
-        : TrackInfo(TrackType::Video) { }
-    bool equalTo(const TrackInfo& otherVideoInfo) const final
-    {
-        auto& other = downcast<const VideoInfo>(otherVideoInfo);
-        return size == other.size && displaySize == other.displaySize && bitDepth == other.bitDepth && colorSpace == other.colorSpace && ((!atomData && !other.atomData) || (atomData && other.atomData && *atomData == *other.atomData));
-    }
-};
-
-struct AudioInfo : public TrackInfo {
-    static Ref<AudioInfo> create() { return adoptRef(*new AudioInfo()); }
-
-    uint32_t rate { 0 };
-    uint32_t channels { 0 };
-    uint32_t framesPerPacket { 0 };
-    uint8_t bitDepth { 16 };
-
-    RefPtr<SharedBuffer> cookieData;
-
-private:
-    AudioInfo()
-        : TrackInfo(TrackType::Audio) { }
-    bool equalTo(const TrackInfo& otherAudioInfo) const final
-    {
-        auto& other = downcast<const AudioInfo>(otherAudioInfo);
-        return rate == other.rate && channels == other.channels && bitDepth == other.bitDepth && framesPerPacket == other.framesPerPacket && ((!cookieData && !other.cookieData) || (cookieData && other.cookieData && *cookieData == *other.cookieData));
-    }
-};
-
-class MediaSamplesBlock {
-    WTF_DEPRECATED_MAKE_FAST_ALLOCATED(MediaSamplesBlock);
-public:
-    struct MediaSampleItem {
-        using MediaSampleDataType = RefPtr<FragmentedSharedBuffer>;
-        MediaTime presentationTime;
-        MediaTime decodeTime { MediaTime::indefiniteTime() };
-        MediaTime duration { MediaTime::zeroTime() };
-        std::pair<MediaTime, MediaTime> trimInterval { MediaTime::zeroTime(), MediaTime::zeroTime() };
-        MediaSampleDataType data;
-        RefPtr<SharedBuffer> hdrMetadata { nullptr };
-        std::optional<HdrMetadataType> hdrMetadataType { std::nullopt };
-        uint32_t flags { };
-        bool isSync() const { return flags & MediaSample::IsSync; }
-    };
-
-    using MediaSampleDataType = MediaSampleItem::MediaSampleDataType;
-    using SamplesVector = Vector<MediaSampleItem>;
-
-    MediaSamplesBlock() = default;
-    MediaSamplesBlock(const TrackInfo* info, SamplesVector&& items)
-        : m_info(info)
-        , m_samples(WTFMove(items))
-    {
-    }
-
-    void setInfo(RefPtr<const TrackInfo>&& info) { m_info = WTFMove(info); }
-    const TrackInfo* info() const { return m_info.get(); }
-    RefPtr<const TrackInfo> protectedInfo() const { return m_info; }
-    MediaTime presentationTime() const { return isEmpty() ? MediaTime::invalidTime() : first().presentationTime; }
-    MediaTime duration() const
-    {
-        MediaTime duration = MediaTime::zeroTime();
-        for (auto& sample : *this)
-            duration += sample.duration;
-        return duration;
-    }
-    MediaTime presentationEndTime() const { return presentationTime() + duration(); }
-    bool isSync() const { return size() ? (first().flags & MediaSample::IsSync) : false; }
-    TrackID trackID() const { return m_info ? m_info->trackID : -1; }
-    bool isVideo() const { return m_info && m_info->isVideo(); }
-    bool isAudio() const { return m_info && m_info->isAudio(); }
-    TrackInfo::TrackType type() const { return m_info ? m_info->type() : TrackInfo::TrackType::Unknown; }
-    void append(MediaSampleItem&& item) { m_samples.append(WTFMove(item)); }
-    void append(MediaSamplesBlock&& block) { append(std::exchange(block.m_samples, { })); }
-    void append(SamplesVector&& samples) { m_samples.appendVector(WTFMove(samples)); }
-    size_t size() const { return m_samples.size(); };
-    bool isEmpty() const { return m_samples.isEmpty(); }
-    void clear() { m_samples.clear(); }
-    SamplesVector takeSamples() { return std::exchange(m_samples, { }); }
-
-    // Indicate that this MediaSampleBlock follows a discontinuity from the previous block.
-    std::optional<bool> discontinuity() const { return m_discontinuity; }
-    void setDiscontinuity(bool discontinuity) { m_discontinuity = discontinuity; }
-
-    const MediaSampleItem& operator[](size_t index) const LIFETIME_BOUND { return m_samples[index]; }
-    const MediaSampleItem& first() const LIFETIME_BOUND { return m_samples.first(); }
-    const MediaSampleItem& last() const LIFETIME_BOUND { return m_samples.last(); }
-    SamplesVector::const_iterator begin() const LIFETIME_BOUND { return m_samples.begin(); }
-    SamplesVector::const_iterator end() const LIFETIME_BOUND { return m_samples.end(); }
-
-private:
-    RefPtr<const TrackInfo> m_info;
-    SamplesVector m_samples;
-    std::optional<bool> m_discontinuity;
-};
-
 } // namespace WebCore
-
-SPECIALIZE_TYPE_TRAITS_BEGIN(WebCore::VideoInfo)
-    static bool isType(const WebCore::TrackInfo& info) { return info.isVideo(); }
-SPECIALIZE_TYPE_TRAITS_END()
-
-SPECIALIZE_TYPE_TRAITS_BEGIN(WebCore::AudioInfo)
-    static bool isType(const WebCore::TrackInfo& info) { return info.isAudio(); }
-SPECIALIZE_TYPE_TRAITS_END()
 
 namespace WTF {
