+++ Source/WebCore/platform/audio/cocoa/AudioDecoderCocoa.cpp
@@ -31,6 +31,7 @@
 #include "AudioSampleBufferConverter.h"
 #include "AudioSampleFormat.h"
 #include "CAAudioStreamDescription.h"
+#include "Logging.h"
 #include "MediaSampleAVFObjC.h"
 #include "MediaUtilities.h"
 #include "PlatformRawAudioData.h"
