+++ Source/WebCore/platform/graphics/cocoa/AudioTrackPrivateWebM.cpp
@@ -29,7 +29,7 @@
 
 #if ENABLE(MEDIA_SOURCE)
 
-#include "MediaSample.h"
+#include "TrackInfo.h"
 
 namespace WebCore {
