+++ Source/WebCore/platform/graphics/cocoa/SourceBufferParserWebM.h
@@ -30,6 +30,7 @@
 #include <CoreAudio/CoreAudioTypes.h>
 #include <WebCore/LibWebRTCMacros.h>
 #include <WebCore/MediaSample.h>
+#include <WebCore/MediaSamplesBlock.h>
 #include <WebCore/SharedBuffer.h>
 #include <WebCore/SourceBufferParser.h>
 #include <pal/spi/cf/CoreMediaSPI.h>
